dt is now in float and in seconds

+++ src/ntshengn_physics_module.cpp
@@ -8,7 +8,7 @@ void NtshEngn::PhysicsModule::init() {
 	NTSHENGN_MODULE_FUNCTION_NOT_IMPLEMENTED();
 }
 
-void NtshEngn::PhysicsModule::update(double dt) {
+void NtshEngn::PhysicsModule::update(float dt) {
 	NTSHENGN_UNUSED(dt);
 	NTSHENGN_MODULE_FUNCTION_NOT_IMPLEMENTED();
 }
